location_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
redis==5.0.7
asyncpg==0.29.0
PyJWT==2.9.0
python-decouple==3.8
